fixup! Fix jmap request to server is called when query string is empty in quick search web
