+++ lib/features/login/presentation/widgets/dns_lookup_input_form.dart
@@ -1,5 +1,5 @@
 
-import 'package:core/presentation/views/quick_search/quick_search_input_form.dart';
+import 'package:core/presentation/views/quick_search/quick_search_action_define.dart';
 import 'package:core/presentation/views/text/type_ahead_form_field_builder.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:tmail_ui_user/features/base/widget/recent_item_tile_widget.dart';
