+++ lib/features/mailbox_dashboard/presentation/widgets/search_input_form_widget.dart
@@ -4,6 +4,9 @@ import 'package:core/presentation/resources/image_paths.dart';
 import 'package:core/presentation/utils/responsive_utils.dart';
 import 'package:core/presentation/views/button/icon_button_web.dart';
 import 'package:core/presentation/views/quick_search/quick_search_input_form.dart';
+import 'package:core/presentation/views/quick_search/quick_search_suggestion_box_decoration.dart';
+import 'package:core/presentation/views/quick_search/quick_search_suggestion_list.dart';
+import 'package:core/presentation/views/quick_search/quick_search_text_field_configuration.dart';
 import 'package:core/utils/app_logger.dart';
 import 'package:core/utils/direction_utils.dart';
 import 'package:flutter/material.dart';
@@ -234,7 +237,7 @@ class SearchInputFormWidget extends StatelessWidget with AppLoaderMixin {
   Widget buildListButtonForQuickSearchForm(
     BuildContext context,
     QuickSearchFilter searchFilter,
-    SuggestionsListState suggestionsListState
+    QuickSearchSuggestionListState suggestionsListState
   ) {
     return Obx(() {
       final isSelected = searchFilter.isApplied(_searchController.listFilterOnSuggestionForm);
